refactor: split protocol and backends from the provider
