+++ merino/configs/default.toml
@@ -104,8 +104,12 @@ top_picks_file_path = "dev/top_picks.json"
 enabled_by_default = false
 # The backend of the provider. Either "elasticsearch" or "test".
 backend = "elasticsearch"
-# Elasticsearch Cloud ID.
-es_cloud_id = ""
+# Elasticsearch Cloud ID. The following one is *fake* and will be
+# overridden in production.
+es_cloud_id = """
+  testing:dXMtZWFzdC0xLmF3cy5mb3VuZC5pbyRjZWM2ZjI2MWE3NGJmMjRjZ
+  TMzYmI4ODExYjg0Mjk0ZiRjNmMyY2E2ZDA0MjI0OWFmMGNjN2Q3YTllOTYyNTc0Mw==
+"""
 # Elasticsearch user
 es_user = ""
 # Elasticsearch password
