+++ merino/providers/wikipedia/__init__.py
@@ -1,0 +1,1 @@
+"""Dynamic Wikipedia provider"""
