+++ merino/providers/wikipedia/backends/__init__.py
@@ -1,0 +1,1 @@
+"""Dynamic Wikipedia backends"""
